requirements.txt
fastapi
uvicorn
gradio
groq
pydantic
python-dotenv
requests
